Move payloads to same place
