+++ PAYLOADS/payload.c
@@ -6,15 +6,15 @@
 
 #define min(a, b) (((a) < (b)) ? (a) : (b))
 
-#define pointToIFO ((void (*)(unsigned int index, unsigned int lba, unsigned int offset))0x23dfc8)
-#define getDiscData ((void (*)(unsigned int s, void *d))0x23e138)
+void (*pointToIFO)(unsigned int index, unsigned int lba, unsigned int offset);
+void (*getDiscData)(unsigned int s, void *d);
 
-#define SifIopReset ((int (*)(char *, int))0x291358)
-#define SifIopSync ((int (*)(void))0x2914d8)
-#define SifInitRpc ((void (*)(int))0x208260)
-#define SifExitRpc ((void (*)(void))0x208400)
+int (*getBufferInternal)(void *filename, int type, int currentSector, void *dest, unsigned int sectorsRemaining, int curReadPos);
 
-#define getBufferInternal ((int (*)(void *filename, int type, int currentSector, void *dest, unsigned int sectorsRemaining, int curReadPos))0x261548)
+int (*SifIopReset)(char *, int);
+int (*SifIopSync)(void);
+void (*SifInitRpc)(int);
+void (*SifExitRpc)(void);
 
 #define ELF_PT_LOAD 1
 
@@ -55,13 +55,14 @@ __attribute__((noreturn)) void ExecPS2(void *entry, void *gp, int argc, char **a
 	//__builtin_unreachable();
 }
 
-void *memcpy_(void *dest, void *src, size_t n) { 
+static void *memcpy_(void *dest, void *src, size_t n) { 
    int i;
    for(i = 0; i < n; i++) ((unsigned char *)dest)[i] = ((unsigned char *)src)[i];
    return dest;
 }
 
-void *memset(void *dest, int c, size_t n) {
+// Todo: maybe cache last sector to save 1 or 2 reads
+static void *memset(void *dest, int c, size_t n) {
 	int i;
 	for(i = 0; i < n; i++) ((unsigned char *)dest)[i] = c;
 	return dest;
@@ -91,10 +92,57 @@ static void readData(void *dest, unsigned int offset, size_t n) {
 }
 
 __attribute__((noreturn)) void _start(void) {
-	//Exit(0);
-	//asm volatile("la $v1, 0x04; la $a0, 0; syscall 0x04");
-
 	int i;
+
+	// Identify version based on jump target location
+	if((*(void **)0x928D24) == (void *)0x15ea540) {
+		// 3.03
+		pointToIFO = (void *)0x2432c8;
+		getDiscData = (void *)0x243438;
+
+		getBufferInternal = (void *)0x262360;
+
+		SifIopReset = (void *)0x291fb8;
+		SifIopSync = (void *)0x292138;
+		SifInitRpc = (void *)0x2082a0;
+		SifExitRpc = (void *)0x208440;
+	}
+	else if((*(void **)0x6D9C3C) == (void *)0x126b7e0) {
+		// 3.04J
+		pointToIFO = (void *)0x23dfe0;
+		getDiscData = (void *)0x23e150;
+
+		getBufferInternal = (void *)0x261560;
+
+		SifIopReset = (void *)0x84fe0;
+		SifIopSync = (void *)0x85110;
+		SifInitRpc = (void *)0x84180;
+		SifExitRpc = (void *)0x84310;
+	}
+	else if((*(void **)0x5f1f38) == (void *)0x1500014) {
+		// 3.10
+		pointToIFO = (void *)0x25c880;
+		getDiscData = (void *)0x25c9f0;
+		
+		getBufferInternal = (void *)0x002986a0;
+
+		SifIopReset = (void *)0x84fe0;
+		SifIopSync = (void *)0x85110;
+		SifInitRpc = (void *)0x84180;
+		SifExitRpc = (void *)0x84310;
+	}
+	else if((*(void **)0x3EA438) == (void *)0x1500014) {
+		// 3.11
+		pointToIFO = (void *)0x258a28;
+		getDiscData = (void *) 0x258b98;
+
+		getBufferInternal = (void *)0x2952f0;
+
+		SifIopReset = (void *)0x20e7d8;
+		SifIopSync = (void *)0x20e958;
+		SifInitRpc = (void *)0x208d80;
+		SifExitRpc = (void *)0x208f20;
+	}
 
 	#ifdef LOAD_FROM_VTS_02_0_IFO
 		// point to VTS_02_0.IFO
@@ -137,5 +185,6 @@ __attribute__((noreturn)) void _start(void) {
 	SifInitRpc(0);
 	SifExitRpc();
 
-	ExecPS2((void *)eh.entry, 0, 0, 0);
+	char *argv[] = { "cdrom0:\\VIDEO_TS\\VTS_02_0.IFO" };
+	ExecPS2((void *)eh.entry, 0, 1, &argv);
 }
